07_AUTOMATION_자동화/playwright_worker/requirements.txt
playwright==1.44.0
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
google-cloud-storage==2.16.0
